added sine function + tests
